+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Added titles method to `ElementEnumeratorBase` and `ElementBase` (minor)
 * Added a file for baking composite chapters called (`bake_composite_chapters`) and the respective spec.
 (minor change)
+* Added `BakeProblemFirstElements` direction (minor)
 
 ## [3.0.0] - 2021-03-17